+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ### v1.0.0.1 Alpha
  - Added method timeline.compose_post
  - Added method timeline.like_post
+ - Added method timeline.get_post